chrome/test/data/xr/e2e_test_files/html/test_non_immersive_stops_during_immersive.html

<!doctype html>
<!--
Tests that a non-immersive session's rAF stops firing when an immersive session
is active, but resumes afterwards.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script src="../resources/webxr_boilerplate.js"></script>
    <script>
      setup({single_test: true});

      let counter = 0;
      let nonImmersiveDuringImmersiveCounter = 0;

      function stepBeforeImmersive() {
        onMagicWindowXRFrameCallback = function() {
          // Verify that we call a rAF once, then make sure any subsequent calls
          // are not done while there is an immersive session.
          onMagicWindowXRFrameCallback = function() {
            if (sessionInfos[sessionTypes.IMMERSIVE].currentSession !== null) {
              // TODO(https://crbug.com/1116261): Investigate a proper product
              // fix so that this can be removed.
              // The inline rAF loop should not be started back again until we
              // have been notified that the immersive session ended; however,
              // a non-immersive rAF may have already been scheduled and not be
              // cancellable once the immersive session returns, leading to a
              // race. To this end, we will allow for at most one non-immersive
              // frame to be called while we have an immersive session.
              nonImmersiveDuringImmersiveCounter++;
              assert_less_than_equal(nonImmersiveDuringImmersiveCounter, 1,
                  "Non-immersive rAF called during immersive session more than allowed");
            }
          }
          finishJavaScriptStep();
        };
      }

      function stepDuringImmersive() {
        // Let the immersive session run for a bit so the non-immersive rAF
        // can fire if it's going to.
        onImmersiveXRFrameCallback = function() {
          if (counter < 10) {
            counter++;
            return;
          }

          onImmersiveXRFrameCallback = null;
          finishJavaScriptStep();
        };
      }

      function stepAfterImmersive() {
        // Make sure we fire at least once after exiting the immersive session
        onMagicWindowXRFrameCallback = function() {
          done();
        };
      }
    </script>
  </body>
</html>
